sdv-blueprint/kuksa-sync/requirements.txt
kuksa-client
grpcio
grpcio-tools
